requirements.txt
streamlit
Pillow
numpy
requests
openai
